+++ sentry-kotlin-multiplatform/src/commonJvmMain/kotlin/io/sentry/kotlin/multiplatform/JvmScopeProvider.kt
@@ -8,7 +8,7 @@ import io.sentry.kotlin.multiplatform.extensions.toKmpUser
 import io.sentry.kotlin.multiplatform.protocol.Breadcrumb
 import io.sentry.kotlin.multiplatform.protocol.User
 
-internal class JvmScopeProvider(private val scope: JvmScope) : ScopeProvider {
+internal class JvmScopeProvider(private val scope: JvmScope) : Scope {
 
     override var level: SentryLevel?
         set(value) {