+++ sentry-kotlin-multiplatform/src/commonJvmTest/kotlin/io/sentry/kotlin/multiplatform/BaseSentryScopeTest.kt
@@ -5,7 +5,6 @@ import io.sentry.SentryOptions
 actual abstract class BaseSentryScopeTest {
     actual fun initializeScope(): Scope {
         val jvmScope = JvmScope(SentryOptions())
-        val jvmScopeProvider = JvmScopeProvider(jvmScope)
-        return Scope(jvmScopeProvider)
+        return JvmScopeProvider(jvmScope)
     }
 }