+++ sentry-samples/kmp-app-mvvm-di/iosApp/AppDelegate.swift
@@ -1,0 +1,24 @@
+import SwiftUI
+import shared
+
+@UIApplicationMain
+class AppDelegate: UIResponder, UIApplicationDelegate {
+
+    var window: UIWindow?
+
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions
+        launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+
+        startKoin()
+        
+        SentrySetupKt.start()
+        
+        let viewController = UIHostingController(rootView: HomeScreen())
+
+        self.window = UIWindow(frame: UIScreen.main.bounds)
+        self.window?.rootViewController = viewController
+        self.window?.makeKeyAndVisible()
+
+        return true
+    }
+}