+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # Changelog
 
-## 0.0.1-alpha.2
+## Unreleased
 
 ### Features
 
 - Improve Objc/Swift experience with @HiddenFromObjc ([#62](https://github.com/getsentry/sentry-kotlin-multiplatform/pull/62))
+- feat: add view hierarchy ([#53](https://github.com/getsentry/sentry-kotlin-multiplatform/pull/53))
 
 ### Fixes