+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## 0.2.1
+
+### Fixes
+
+- fix: beforeBreadcrumb discarding if hook is not set ([#105](https://github.com/getsentry/sentry-kotlin-multiplatform/pull/105))
+
+## 0.2.0
+
+### Features
+
+- feat: automatically disable `io.sentry.auto-init` ([#93](https://github.com/getsentry/sentry-kotlin-multiplatform/pull/93))
+
+### Fixes
+
+- fix: NSNumber to Kotlin Long crash during SentryException conversion ([#92](https://github.com/getsentry/sentry-kotlin-multiplatform/pull/92))
+
+### Improvements
+
+- ref: improve samples & add SPM docs ([#82](https://github.com/getsentry/sentry-kotlin-multiplatform/pull/82))
+
 ## 0.1.1
 
 ### Fixes